chd: cast before shifting in bitstream_peek - signed overflow on ordinary data
    bitstream->buffer |= bitstream->read[bitstream->doffset] << (24 - bitstream->bits);

read[] is a uint8_t. The usual arithmetic conversions promote it to int,
so any byte from 0x80 up shifted left by 24 overflows a signed int, which
is undefined behaviour. 24 is the shift used whenever the accumulator is
empty, so it is the common case rather than an edge one.

Same defect, and same fix, as get_bigendian_uint32_t in libchdr_chd.c.
This one sits in the huffman map decoder's inner refill, so it runs over
every byte of a compressed CHD's map - it fires while opening ordinary,
well-formed compressed images, not only corrupt ones. Behaviour is
unchanged on any two's-complement machine; only the undefined shift goes
away.

Found by fuzzing libchdr under ASan and UBSan. The earlier pass used an
uncompressed CHD v5 seed, which never enters decompress_v5_map at all and
so never touches the bitstream reader or any codec. Rebuilding the corpus
with compressed seeds - one per codec tag: zlib, lzma, huff, flac, zstd,
cdzl, cdlz, cdfl, cdzs - reached this within six iterations, and four
separate reproducers across four different codec seeds all landed on this
line. Coverage of the compressed path was verified directly with a probe
in decompress_v5_map before fuzzing, so the new corpus was known to reach
code the old one did not.

Verified: all four reproducers go from a UBSan abort to a clean exit, and
every compressed seed still opens and decodes unchanged.

+++ libretro-common/formats/libchdr/libchdr_bitstream.c
@@ -52,7 +52,13 @@ uint32_t bitstream_peek(struct bitstream* bitstream, int numbits)
 		while (bitstream->bits <= 24)
 		{
 			if (bitstream->doffset < bitstream->dlength)
-				bitstream->buffer |= bitstream->read[bitstream->doffset] << (24 - bitstream->bits);
+				/* Cast before shifting. read[] is a uint8_t, promoted to
+				 * int, so a byte from 0x80 up shifted left by 24 - which is
+				 * the shift used whenever the accumulator is empty -
+				 * overflows a signed int. This is the huffman map decoder's
+				 * inner loop, so it runs over every byte of a compressed
+				 * CHD's map. */
+				bitstream->buffer |= (uint32_t)bitstream->read[bitstream->doffset] << (24 - bitstream->bits);
 			bitstream->doffset++;
 			bitstream->bits += 8;
 		}
